Changes geometry methods from float to double
Knocks on to some other classes that depend on geometry information.

+++ StRoot/StFmsPointMaker/StFmsClusterFitter.cxx
@@ -48,7 +48,7 @@ std::vector<double> defaultMinuitStepSizes() {
   }  // for
   return steps;
 }
-std::vector<float> towerWidths;  // Tower (x, y) width in cm
+std::vector<double> towerWidths;  // Tower (x, y) width in cm
 
 // Helper function for accumulating tower energies
 double addTowerEnergy(double energy, const FMSCluster::StFmsTower* tower) {
